requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
opensearch-py==2.3.2
google-cloud-vision==3.4.4
openai==1.3.5
pydantic==2.5.2
<<<<<<< HEAD
python-jose==3.3.0
requests==2.31.0
pillow==10.1.0
numpy==1.26.2 
=======
pydantic-settings==2.1.0
python-jose==3.3.0
requests==2.31.0
pillow==10.1.0
numpy==1.26.2
aiohttp==3.9.1
>>>>>>> dev
